Refactor Material Editor UI architecture
Preserve the existing public API while separating edit operations, storage repositories, and row rendering. Add label click extension events, API compatibility checks, and collapsible shader property categories.
