+++ src/MaterialEditor.Base/MaterialEditor.Base.projitems
@@ -12,6 +12,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)NormalMapManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IMaterialEditorColorPalette.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MaterialAPI.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)MaterialEditService.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ConfigurationManagerAttributes.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CopyContainer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Export.Obj.cs" />
@@ -23,9 +24,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.AutoScrollToCenter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.FloatLabelDragTrigger.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.DropdownFilter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.ItemInfo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.ItemTemplate.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.ListEntry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.RowModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.RowViewFactory.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.LabelClick.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.RowBinder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\UI.RowView.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.Tooltip.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.SelectListPanel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.TooltipManager.cs" />
