+++ src/MaterialEditor.Core.Studio/Core.MaterialEditor.Studio.projitems
@@ -10,6 +10,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneController.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.SceneRepository.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.Studio.ColorPalette.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.Studio.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.TimelineCompatibilityHelper.cs" />
